tests/wpt/web-platform-tests/css/css-transforms/parsing/transform-box-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 1: parsing transform-box with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-transforms/#transform-box-property">
<meta name="assert" content="transform-box supports the full grammar 'content-box | border-box | fill-box | stroke-box | view-box'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("transform-box", "content-box");
test_valid_value("transform-box", "border-box");
test_valid_value("transform-box", "fill-box");
test_valid_value("transform-box", "stroke-box");
test_valid_value("transform-box", "view-box");
</script>
</body>
</html>
